add stringReplacement acourt data to formular
