+++ app/domains/fluggastrechte/formular/__test__/stringReplacements.test.ts
@@ -17,10 +17,7 @@ import {
   isNichtBefoerderung,
   isVerspaetet,
   isWeiterePersonen,
-  getButtonURLForClaimViaPost,
 } from "../stringReplacements";
-import { hasAirportPartnerCourt } from "~/services/airports/hasPartnerCourt";
-
 describe("stringReplacements", () => {
   describe("getArrayWeiterePersonenIndexStrings", () => {
     it("should return an array weitere personen index for given context", () => {
@@ -373,24 +370,6 @@ describe("stringReplacements", () => {
         endAirport: "",
       });
       expect(actual).toStrictEqual({});
-    });
-  });
-  describe("getButtonURLForClaimViaPost", () => {
-    vi.mock("~/services/airports/hasPartnerCourt");
-    it("should return the link to ergebnis/erfolg when partnerCourt exists", () => {
-      vi.mocked(hasAirportPartnerCourt).mockReturnValue(true);
-      const actual = getButtonURLForClaimViaPost({});
-      expect(actual).toStrictEqual({
-        claimViaPostButtonURL: "/fluggastrechte/vorabcheck/ergebnis/erfolg",
-      });
-    });
-    it("should return the link to ergebnis/erfolg-analog when there is no partnerCourt", () => {
-      vi.mocked(hasAirportPartnerCourt).mockReturnValue(false);
-      const actual = getButtonURLForClaimViaPost({});
-      expect(actual).toStrictEqual({
-        claimViaPostButtonURL:
-          "/fluggastrechte/vorabcheck/ergebnis/erfolg-analog",
-      });
     });
   });
 });
