+++ app/domains/fluggastrechte/formular/index.ts
@@ -26,6 +26,7 @@ import {
   getFirstZwischenstoppAirportName,
   getPersonNachname,
   getPersonVorname,
+  getResponsibleCourt,
   getSecondZwischenstoppAirportName,
   getStartAirportName,
   getThirdZwischenstoppAirportName,
@@ -71,6 +72,7 @@ export const fluggastrechtFlow = {
     ...getFirstZwischenstoppAirportName(context),
     ...getSecondZwischenstoppAirportName(context),
     ...getThirdZwischenstoppAirportName(context),
+    ...getResponsibleCourt(context),
     ...isVerspaetet(context),
     ...isNichtBefoerderung(context),
     ...isAnnullierung(context),
