+++ app/domains/fluggastrechte/vorabcheck/index.ts
@@ -3,6 +3,7 @@ import fluggastrechteVorabcheckFlow from "~/domains/fluggastrechte/vorabcheck/fl
 import { guards as fluggastrechteVorabcheckGuards } from "~/domains/fluggastrechte/vorabcheck/guards";
 import type { FluggastrechtVorabcheckContext } from "./context";
 import {
+  getButtonURLForClaimViaPost,
   getCompensationPaymentString,
   getEndAirportName,
   getLastDayFromFourYearsAgoDate,
@@ -15,10 +16,7 @@ import {
   hasCompensationShortDistance,
 } from "./stringReplacements";
 import type { Context } from "../../contexts";
-import {
-  getButtonURLForClaimViaPost,
-  getResponsibleCourt,
-} from "../formular/stringReplacements";
+import { getResponsibleCourt } from "../formular/stringReplacements";
 
 export const fluggastrechteVorabcheck = {
   flowType: "vorabCheck",
